feat: send image messages
The image data from ImageResult is now uploaded to the server and
attached to the message. The data is never populated, so usage of the
bot will likely result in an error or at the very least no image to be
generated. When image generation is implemented then the data will flow
through to the messages the bot sends.